Improve Period api.
Change Period.parse to take one or two string arguments, the first is a
from-spec, the second the to-spec.  This was the hard to remember
parse_spec. Period.parse('2011', '2012-3Q') would yield the period from
2011-01-01 to 2012-09-30.

Re-implement Period.parse_phrase to return a period by extracting the to
and from specs from a single-string phrase that introduces the parts
with 'from' and 'to'.  Period.parse_phrase('from 2011 to 2012-3Q') would
yield the period from 2011-01-01 to 2012-09-30.