Be more careful to distinguish false from nil and allow nils.
Also allow nils in already-typed column.  Old code was confusing a nil returned
by the convert_to... methods and a nil supplied as a value of the column.  The
latter should be perfectly legitimate to indicate unknown data.  The former
means that the supplied value could not be interpreted as the given type.