Change the convention for marking an aoa row as header; mark boundaries.
We were using the first row of an aoa as a header simply if that row had no
numbers, which seems risky and unduly restrictive. Now, the first row is a
header if the second is a nil or a string that looks like an hrule, i.e.,
something like '|------', or if the first element of that row looks like an
hrule. Also, use the nil or hrule convention to mark the preceding row as a
group boundary.  This breaks any aoa's that relied on the old method of
determining the header.