Add @group instance var for select and where; fix @row.
The expressions in #select and #where may now reference @group which is set to
the group number for each row being evaluated.

I also determined that the @row variable was not working correctly.  I had the
before hook for the evaluator incrementing row by 1, but the before hook gets
called for each hash param to #select, so if there were two hash params, it
would get incremented twice per row.  This is now fixed.  We don't use the
before hook to increment now, but set the row number directly in a private local
variable, __row, and the before hook just sets @row to __row.  The same
mechanism is used for @group.