Make constructors explicit as to expected input.
Instead of taking a generic input from Table.new, require the user to say what
the expected input is.  So now, Table.new just returns an empty Table.  To make
a table from an org file, you must say, Table.from_org_file('file_name'), and so
forth.  I want to add something like Table.from_sql(query), and since query is
just a string, there is no way, short of writing a parser, to determine that the
string is meant as a SQL query instead of, say, a CSV string.  This will allow
adding any number of constructors in the future.