Fix failure of type-based format instructions.
If any instructions were given for a numeric column, say :shares, it would
overwrite all instructions for numeric with the defaults.  The problem was that
the parse_... methods were setting all hash elements to their defaults, even if
the format string contained no corresponding instruction.  So, e.g., setting
numeric to 'R' for right-align and :shares to "0.4," would reset alignment on
shares column to left.