Decorate cells on the first round if the Formatter is no aligned.
Doecorating took place in the second pass over the table (after the widths were
established) but not on the first pass. This causes decoration to occur on the
first pass when the Formatter is not doing its own alignment (e.g.,
LaTeXFormatter) since otherwise, the decorate method never gets called.