Change Date back to monkey-patch form; complete #next_<chunk> methods.
Looked like the documentation gets confusing when I changed Date to a Module
that got included into the Date class.  The class methods included with the
\#included hook in particular did not show up in the documentation for the Date
class.  Also, the Module method polluted the Date name space just as much as
directly patching the Date class, so I don't see the advantage.

Added #next_semimonth, #next_biweek, etc. to have a complete set of
\#next_<chunk> methods.  Also changed the way #next_semimonth is calculated to
take into account where we are in the month.